+++ main/parametertune.cpp
@@ -4,21 +4,21 @@
 
 using namespace std;
 
-double utility(GrowingNetwork3D* net);
+float utility(GrowingNetwork3D* net);
 
 int main(){
 	
 	long int sample_size = 25;
 	long int n = 100, m = 3;
-	double base_gamma_step = 0.01, base_tol_step = 0.01;
+	float base_gamma_step = 0.01, base_tol_step = 0.01;
 	long int base_itr_step = 1;
-	double gamma = 2.8, gamma_step = base_gamma_step, tolerance = 0.6, tol_step = base_tol_step;
+	float gamma = 2.8, gamma_step = base_gamma_step, tolerance = 0.6, tol_step = base_tol_step;
 	long int iterations = 40, itr_step = base_itr_step;
 	long int gen = 0, genmax = 400;
 	GrowingNetwork3D* net = new GrowingNetwork3D(n, m, gamma, tolerance, iterations);
 
-	double prev_util = utility(net);
-	double best_gamma, best_tolerance;
+	float prev_util = utility(net);
+	float best_gamma, best_tolerance;
 	long int best_iterations;
 
 	while(gen < genmax){
@@ -33,10 +33,10 @@ int main(){
 			#pragma omp for schedule(dynamic)
 			for(long int i = 0; i < 8; i++){
 				
-				double test_gamma = abs(gamma + ((i & 1<<0) ? -1:1) * gamma_step);
-				double test_tolerance = abs(tolerance + ((i & 1<<1) ? -1:1) * tol_step);
+				float test_gamma = abs(gamma + ((i & 1<<0) ? -1:1) * gamma_step);
+				float test_tolerance = abs(tolerance + ((i & 1<<1) ? -1:1) * tol_step);
 				long int test_iterations = abs(iterations + ((i & 1<<2) ? -1:1) * itr_step);
-				double test_util = 0;
+				float test_util = 0;
 				
 				for(long int j = 0; j < sample_size; j++){
 					
@@ -101,7 +101,7 @@ int main(){
 	
 }
 
-double utility(GrowingNetwork3D* net){
+float utility(GrowingNetwork3D* net){
 	
 	return net->calculateMinimumPotential(net->N, 3) - net->calculatePotential();
